refactor: extract external events from App
